requirements.txt
faster_whisper==1.1.1
fire==0.7.0
Jinja2==3.1.2
numpy==1.24.3
pandas==1.5.3
python-dotenv==1.0.1
Requests==2.32.3
scikit_learn==1.2.2
torch==2.2.2
tqdm==4.65.0
transformers==4.49.0
